extracts reclassification to a module
